CanopyOptics.jl

A package to compute canopy scattering properties

Package Features

Use leaf angle distributions to compute bi-lambertian area scattering matrices

Compute specular reflection

Compute leaf reflectance and transmittance based on Prospect-PRO

Installation

The latest release of CanopyOptics can be installed from the Julia REPL prompt with

julia> ]add https://github.com/RemoteSensingTools/CanopyOptics.jl

Code docs:

Types

CanopyOptics.AbstractCanopyScatteringType — Type

Abstract Type for canopy scattering

CanopyOptics.BiLambertianCanopyScattering — Type

Model for bi-lambertian canopy leaf scattering

CanopyOptics.LeafOpticalProperties — Type

struct PigmentOpticalProperties{FT}

A struct which stores important absorption cross sections of pigments, water, etc

Fields

λ

Wavelength [length]

nᵣ

Refractive index of leaf material

Kcab

specific absorption coefficient of chlorophyll (a+b) [cm² μg⁻¹]

Kcar

specific absorption coefficient of carotenoids [cm² μg⁻¹]

Kant

specific absorption coefficient of Anthocyanins [cm² nmol⁻¹]

Kb

specific absorption coefficient of brown pigments (arbitrary units)

Kw

specific absorption coefficient of water [cm⁻¹]

Km

specific absorption coefficient of dry matter [cm² g⁻¹]

Kp

specific absorption coefficient of proteins [cm² g⁻¹]

Kcbc

specific absorption coefficient of carbon based constituents [cm² g⁻¹]

CanopyOptics.LeafProspectProProperties — Type

struct LeafProperties{FT}

A struct which stores important variables of leaf chemistry and structure

Fields

N

Leaf structure parameter [0-3]

Ccab

Chlorophyll a+b content [µg cm⁻²]

Ccar

Carotenoid content [µg cm⁻²]

Canth

Anthocynanin content [nmol cm⁻²]

Cbrown

Brown pigments content in arbitrary units

Cw

Equivalent water thickness [cm], typical about 0.002-0.015

Cm

Dry matter content (dry leaf mass per unit area) [g cm⁻²], typical about 0.003-0.016

Cprot

protein content [g/cm]

Ccbc

Carbone-based constituents content in [g/cm⁻²] (cellulose, lignin, sugars...)

CanopyOptics.SpecularCanopyScattering — Type

Model for specular canopy leaf scattering

CanopyOptics.dirVector — Type

Struct for spherical coordinate directions in θ (elevation angle) and ϕ (azimuth angle)

CanopyOptics.dirVector_μ — Type

Struct for spherical coordinate directions in θ (elevation angle) and ϕ (azimuth angle)

Functions

CanopyOptics.createLeafOpticalStruct — Method

createLeafOpticalStruct(λ_bnds)

Loads in the PROSPECT-PRO database of pigments (and other) absorption cross section in leaves, returns a [`LeafOpticalProperties`](@ref) type struct with spectral units attached.

Arguments

- `λ_bnds` an array (with or without a spectral grid unit) that defines the upper and lower limits over which to average the absorption cross sections

Examples

julia> using Unitful                                                               # Need to include Unitful package 
julia> opti = createLeafOpticalStruct((400.0:5:2400)*u"nm");                       # in nm
julia> opti = createLeafOpticalStruct((0.4:0.1:2.4)*u"μm");                        # in μm
julia> opti = CanopyOptics.createLeafOpticalStruct((10000.0:100:25000.0)u"1/cm");  # in wavenumber (cm⁻¹)

CanopyOptics.createLeafOpticalStruct — Method

createLeafOpticalStruct()
As in createLeafOpticalStruct(λ_bnds) but reads in the in Prospect-PRO database at original resolution (400-2500nm in 1nm steps)

CanopyOptics.prospect — Method

prospect(leaf::LeafProspectProProperties{FT},
                optis) where {FT<:AbstractFloat}

Computes leaf optical properties (reflectance and transittance) based on PROSPECT-PRO

Arguments

leaf LeafProspectProProperties type struct which provides leaf composition

optis LeafOpticalProperties type struct, which provides absorption cross sections and spectral grid

Examples

julia-repl julia> opti = createLeafOpticalStruct((400.0:5:2400)*u"nm"); julia> leaf = LeafProspectProProperties{Float64}(Ccab=30.0); julia> T,R = prospect(leaf,opti);`