CanopyOptics.jl

A package to compute canopy scattering properties

Package Features

Use leaf angle distributions to compute bi-lambertian area scattering matrices

Compute specular reflection

Compute leaf reflectance and transmittance based on Prospect-PRO

Installation

The latest release of CanopyOptics can be installed from the Julia REPL prompt with

julia> ]add https://github.com/RemoteSensingTools/CanopyOptics.jl

Function Documentation

CanopyOptics.AbstractCanopyScatteringType — Type

Abstract Type for canopy scattering

CanopyOptics.AbstractLeafDistribution — Type

Abstract Type for leaf distributions

CanopyOptics.AbstractLeafProperties — Type

Abstract Type for leaf composition

CanopyOptics.AbstractProspectProperties — Type

Abstract type for Prospect model versions

CanopyOptics.BiLambertianCanopyScattering — Type

Model for bi-lambertian canopy leaf scattering

CanopyOptics.LeafDistribution — Type

struct LeafDistribution{FT<:AbstractFloat}

A struct that defines the leaf angular distribution in radians (from 0->π/2; here scaled to [0,1])

Fields

LD

Julia Univariate Distribution from Distributions.js

scaling

Scaling factor to normalize distribution (here mostly 2/π as Beta distribution is from [0,1])

CanopyOptics.LeafOpticalProperties — Type

struct PigmentOpticalProperties{FT}

A struct which stores important absorption cross sections of pigments, water, etc

Fields

λ

Wavelength [length]

nᵣ

Refractive index of leaf material

Kcab

specific absorption coefficient of chlorophyll (a+b) [cm² μg⁻¹]

Kcar

specific absorption coefficient of carotenoids [cm² μg⁻¹]

Kant

specific absorption coefficient of Anthocyanins [cm² nmol⁻¹]

Kb

specific absorption coefficient of brown pigments (arbitrary units)

Kw

specific absorption coefficient of water [cm⁻¹]

Km

specific absorption coefficient of dry matter [cm² g⁻¹]

Kp

specific absorption coefficient of proteins [cm² g⁻¹]

Kcbc

specific absorption coefficient of carbon based constituents [cm² g⁻¹]

CanopyOptics.LeafProspectProProperties — Type

struct LeafProperties{FT}

A struct which stores important variables of leaf chemistry and structure

Fields

N

Leaf structure parameter [0-3]

Ccab

Chlorophyll a+b content [µg cm⁻²]

Ccar

Carotenoid content [µg cm⁻²]

Canth

Anthocynanin content [nmol cm⁻²]

Cbrown

Brown pigments content in arbitrary units

Cw

Equivalent water thickness [cm], typical about 0.002-0.015

Cm

Dry matter content (dry leaf mass per unit area) [g cm⁻²], typical about 0.003-0.016

Cprot

protein content [g/cm]

Ccbc

Carbone-based constituents content in [g/cm⁻²] (cellulose, lignin, sugars...)

CanopyOptics.SpecularCanopyScattering — Type

Model for specular canopy leaf scattering

CanopyOptics.dirVector — Type

Struct for spherical coordinate directions in θ (elevation angle) and ϕ (azimuth angle)

CanopyOptics.dirVector_μ — Type

Struct for spherical coordinate directions in θ (elevation angle) and ϕ (azimuth angle)

CanopyOptics.A — Method

Integrated projection of leaf area for a single leaf inclination of θₗ, assumes azimuthally uniform distribution

CanopyOptics.Fᵣ — Method

Fresnel reflection, needs pol_type soon too!

CanopyOptics.G — Method

G(μ::Array{FT}, LD::AbstractLeafDistribution; nLeg=20)

Returns the integrated projection of leaf area in the direction of μ, assumes azimuthally uniform distribution and a LD distribution for leaf polar angle θ. This function is often referred to as the
function O(B) (Goudriaan 1977) or G(Ζ) (Ross 1975,1981), see Bonan modeling book, eqs. 14.21-14.26.

Arguments

μ an array of cos(θ) (directions [0,1])

LD an AbstractLeafDistribution type struct, includes a leaf distribution function

nLeg an optional parameter for the number of legendre polynomials to integrate over the leaf distribution (default=20)

Examples

julia> μ,w = CanopyOptics.gauleg(10,0.0,1.0);       # Create 10 quadrature points in μ      
julia> LD  = CanopyOptics.spherical_leaves()        # Create a default spherical leaf distribution
julia> G   = CanopyOptics.G(μ, LD)                  # Compute G(μ)
10-element Vector{Float64}:
 0.5002522783000879
 0.5002715115149204
 0.5003537989277846
 0.5004432798701134
 0.5005134448870893
 0.5003026448466977
 0.4999186257540982
 0.4994511190721635
 0.49907252201082375
 0.49936166823681594

CanopyOptics.H — Method

Eq 46 in Shultis and Myneni

CanopyOptics.K — Method

The reduction factor proposed by Nilson and Kuusk, κ ≈ 0.1-0.3

CanopyOptics.calctav — Method

calctav(α::FT, nr::FT) where {FT<:AbstractFloat}

Computes transmission of isotropic radiation across a dielectric surface (Stern F., 1964; Allen W.A.,Appl. Opt., 3(1):111-113 1973)). From calctav.m in PROSPECT-D

Arguments

α angle of incidence [degrees]

nr Index of refraction

CanopyOptics.compute_Z_matrices — Method

compute_Z_matrices(mod::BiLambertianCanopyScattering, μ::Array{FT,1}, LD::AbstractLeafDistribution, m::Int)

Computes the single scattering Z matrices (𝐙⁺⁺ for same incoming and outgoing sign of μ, 𝐙⁻⁺ for a change in direction). Internally computes the azimuthally-averaged area scattering transfer function following Shultis and Myneni (https://doi.org/10.1016/0022-4073(88)90079-9), Eq 43::

$Γ(μ' -> μ) = \int_0^1 dμ_L g_L(μ_L)[t_L Ψ⁺(μ, μ', μ_L) + r_L Ψ⁻(μ, μ', μ_L)]$

assuming an azimuthally uniform leaf angle distribution. Normalized Γ as 𝐙 = 4Γ/(ϖ⋅G(μ)). Returns 𝐙⁺⁺, 𝐙⁻⁺

Arguments

mod : A bilambertian canopy scattering model BiLambertianCanopyScattering, uses R,T,nQuad from that model.

μꜛ::Array{FT,1}: Quadrature points ∈ [0,1]

LD a AbstractLeafDistribution struct that describes the leaf angular distribution function.

m: Fourier moment (for azimuthally uniform leave distributions such as here, only m=0 returns non-zero matrices)

CanopyOptics.compute_lambertian_Γ — Method

compute_lambertian_Γ(μ::Array{FT,1},μꜛ::Array{FT,1}, r,t, LD::AbstractLeafDistribution; nLeg = 20)

Computes the azimuthally-averaged area scattering transfer function following Shultis and Myneni (https://doi.org/10.1016/0022-4073(88)90079-9), Eq 43:

$Γ(μ' -> μ) = \int_0^1 dμ_L g_L(μ_L)[t_L Ψ⁺(μ, μ', μ_L) + r_L Ψ⁻(μ, μ', μ_L)]$

assuming an azimuthally uniform leaf angle distribution.

Arguments

μ::Array{FT,1} : Quadrature points incoming direction (cos(θ))

μꜛ::Array{FT,1}: Quadrature points outgoing direction (cos(θ))

r : Leaf lambertian reflectance

t : Leaf lambertian transmittance

LD a AbstractLeafDistribution struct that describes the leaf angular distribution function.

nLeg = 20: number of quadrature points used for integration over all leaf angles (default is 20).

CanopyOptics.compute_Ψ — Method

Eq 45 in Shultis and Myneni, fixed grid in μ for both μ and μ'

CanopyOptics.compute_Ψ — Method

Eq 45 in Shultis and Myneni, fixed grid in μ for both μ and μ'

CanopyOptics.createLeafOpticalStruct — Method

createLeafOpticalStruct(λ_bnds)

Loads in the PROSPECT-PRO database of pigments (and other) absorption cross section in leaves, returns a [`LeafOpticalProperties`](@ref) type struct with spectral units attached.

Arguments

- `λ_bnds` an array (with or without a spectral grid unit) that defines the upper and lower limits over which to average the absorption cross sections

Examples

julia> using Unitful                                                               # Need to include Unitful package 
julia> opti = createLeafOpticalStruct((400.0:5:2400)*u"nm");                       # in nm
julia> opti = createLeafOpticalStruct((0.4:0.1:2.4)*u"μm");                        # in μm
julia> opti = CanopyOptics.createLeafOpticalStruct((10000.0:100:25000.0)u"1/cm");  # in wavenumber (cm⁻¹)

CanopyOptics.createLeafOpticalStruct — Method

createLeafOpticalStruct()
As in createLeafOpticalStruct(λ_bnds) but reads in the in Prospect-PRO database at original resolution (400-2500nm in 1nm steps)

CanopyOptics.erectophile_leaves — Function

erectophile_leaves(FT=Float64)

Creates a LeafDistribution following a erectophile (mostly vertical) distribution using a Beta distribution

Standard values for θₗ (63.24) and s (18.5) from Bonan https://doi.org/10.1017/9781107339217.003, Table 2.1

CanopyOptics.plagiophile_leaves — Function

plagiophile_leaves(FT=Float64)

Creates a LeafDistribution following a plagiophile (mostly 45degrees) distribution using a Beta distribution

Standard values for θₗ (45.0) and s (16.27) from Bonan https://doi.org/10.1017/9781107339217.003, Table 2.1

CanopyOptics.planophile_leaves — Function

planophile_leaves(FT=Float64)

Creates a LeafDistribution following a planophile (mostly horizontal) distribution using a Beta distribution

Standard values for θₗ (26.76) and s (18.51) from Bonan https://doi.org/10.1017/9781107339217.003, Table 2.1

CanopyOptics.prospect — Method

prospect(leaf::LeafProspectProProperties{FT},
                optis) where {FT<:AbstractFloat}

Computes leaf optical properties (reflectance and transittance) based on PROSPECT-PRO

Arguments

leaf LeafProspectProProperties type struct which provides leaf composition

optis LeafOpticalProperties type struct, which provides absorption cross sections and spectral grid

Examples

julia-repl julia> opti = createLeafOpticalStruct((400.0:5:2400)*u"nm"); julia> leaf = LeafProspectProProperties{Float64}(Ccab=30.0); julia> T,R = prospect(leaf,opti);`

CanopyOptics.scattering_model_lambertian — Method

Eq 37 in Shultis and Myneni

CanopyOptics.spherical_leaves — Function

spherical_leaves(FT=Float64)

Creates a LeafDistribution following a spherical (distributed as facets on a sphere) distribution using a Beta distribution

Standard values for θₗ (57.3) and s (21.55) from Bonan https://doi.org/10.1017/9781107339217.003, Table 2.1

CanopyOptics.uniform_leaves — Function

uniform_leaves(FT=Float64)

Creates a LeafDistribution following a uniform distribution (all angles equally likely)

CanopyOptics.unit2nm — Method

Convert an input array with Spectral units (or none) to a grid in nm

CanopyOptics.βparameters — Method

Get Beta Distribution parameters using standard tabulated values θₗ,s, Eq. 2.9 and 2.10 in Bonan